Exporter

+++ src/app/services/cv-parser.service.ts
@@ -1,10 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { createClient } from '@supabase/supabase-js';
-import { AppUser, CV } from '../model/user.model';
+import { AppUser, CV,Competence,ParsedResume, Resume } from '../model/user.model';
 import { Router } from '@angular/router';
 import { SupabaseClientService } from './supabase-client.service';
-import * as pdfjsLib from 'pdfjs-dist';
+
 
 @Injectable({
   providedIn: 'root'
@@ -19,8 +19,7 @@ export class CvParserService {
     this.supabaseUrl = 'https://mljtanxsvdnervhrjnbs.supabase.co';
     this.supabaseKey = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im1sanRhbnhzdmRuZXJ2aHJqbmJzIiwicm9sZSI6ImFub24iLCJpYXQiOjE2ODQ4NDczMDQsImV4cCI6MjAwMDQyMzMwNH0.lrhe---iFdN9RSFGgF5cYwN9S_aWpxYGur1TAvrD-ZY';
     this.supabase = createClient(this.supabaseUrl, this.supabaseKey);
-    const workerPath = 'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.16.105/pdf.worker.min.js';
-    pdfjsLib.GlobalWorkerOptions.workerSrc = workerPath;
+
    }
 
   async addCV(cvDetails: any) {
@@ -54,150 +53,15 @@ export class CvParserService {
       reader.onerror = (error) => reject(error);
     });
   }
-  async parseResume(base64File: string): Promise<any> {
-    console.log('Parsing resume with base64 data:', base64File);
-    const pdfData = this.convertBase64ToUint8Array(base64File);    
-    // Extract text from the PDF
-    const extractedText = await this.extractTextFromPDF(pdfData);
-    console.log('Extracted text:', extractedText);
-      console.log('JSON stringified text:', JSON.stringify(extractedText));
-    // Extract candidate's name using regex
-    const nameRegex = /[A-Z][a-zÀ-ÿ']*(?:\s[A-Z][a-zÀ-ÿ']*)+/;
-    const extractedNames = extractedText.match(nameRegex);
-    let candidateName = "Nom du candidat inconnu";
-    if (extractedNames && extractedNames.length > 0) {
-      candidateName = extractedNames[0];
-      console.log(`Candidate Name: ${candidateName}`);
-    }
-
-    // Extract candidate's email using regex
-    const candidateEmailRegex = /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}/i;
-    const matchEmail = extractedText.match(candidateEmailRegex);
-    const candidateEmail = matchEmail ? matchEmail[0] : "Adresse e-mail inconnue";
-    console.log("Candidate Email:", candidateEmail);
-  
-    // Extract candidate's phone number using regex
-    const candidateNumRegex = /(?:0[67]\d{8})|(?:\+\d+)/i;
-    const candidateNumMatch = extractedText.match(candidateNumRegex);
-    const candidateNum = candidateNumMatch ? candidateNumMatch[0] : "Num inconnu";
-    console.log("Candidate Num:", candidateNum);
-    // Extract candidate's postal address using regex
-    const addressRegex = /ADRESSE:\s*([\s\S]*?)/;
-    const addressMatch = extractedText.match(addressRegex);
-    console.log('Address match:', addressMatch);
-    let postalAddress = "Adresse inconnue";
-    if (addressMatch) {
-      postalAddress = addressMatch[1];
-      console.log('Postal Address:', postalAddress);
-    } else {
-      console.log('No postal address found.');
-    }
-    // Extract candidate's educational background using regex
-  const educationRegex = /FORMATION(?:S)?:\s*([\s\S]*?)(?=COMPÉTENCES:|$)/;
-  const educationMatch = extractedText.match(educationRegex);
-  console.log('Education match:', educationMatch);
-  
-  let educationBackground = "Formation inconnue";
-  if (educationMatch && educationMatch[1]) {
-    educationBackground = educationMatch[1]
-      .split(/[\r\n]+/) 
-      .map(line => line.trim()) 
-      .filter(line => line.length > 0) 
-      .join('\n'); 
-    console.log('Education Background:', educationBackground);
-  } else {
-    console.log('No education background found.');
-  }
-
-    // Extract candidate's competences using regex
-    const competenceRegex = /COMPÉTENCES:\s*([\s\S]*?)(?=QUALITÉS PERSONNELLES:|$)/;
-    const competenceMatch = extractedText.match(competenceRegex);
-    console.log('Competence match:', competenceMatch);
-    
-    let competences = "Compétences inconnues";
-    if (competenceMatch && competenceMatch[1]) {
-      competences = competenceMatch[1]
-        .split(/[\r\n]+/) 
-        .map(line => line.trim()) 
-        .filter(line => line.length > 0) 
-        .join(', '); 
-        console.log('Compétences:', competences);
-    }else{
-      console.log('Aucune compétence trouvée.');
-    }
-  // Extract candidate's certifications using regex
-    const certificationRegex = /CERTIFICATIONS:\s*([\s\S]*?)(?=(LANGUES:|$))/;
-    const certificationMatch = extractedText.match(certificationRegex);
-    console.log('Certification match:', certificationMatch);
-
-    let certifications = "Aucune certification trouvée";
-    if (certificationMatch && certificationMatch[1]) {
-      certifications = certificationMatch[1]
-        .split(/[\r\n]+/) 
-        .map(line => line.trim()) 
-        .filter(line => line.length > 0) 
-        .join('\n'); 
-      console.log('Certifications:', certifications);
-    } else {
-      console.log('No certifications found.');
-    }
-    // Extract candidate's skills using regex
-    const skillsRegex = /COMPÉTENCES:\s*([\s\S]*?)(?=(INTÉRÊTS:|$))/;
-    const skillsMatch = extractedText.match(skillsRegex);
-    console.log('Skills match:', skillsMatch);
-
-    let skills = "Aucune compétence trouvée";
-    if (skillsMatch && skillsMatch[1]) {
-      skills = skillsMatch[1]
-        .split(/[,;]+/) 
-        .map(skill => skill.trim()) 
-        .filter(skill => skill.length > 0) 
-        .join(', '); 
-      console.log('Skills:', skills);
-    } else {
-      console.log('No skills found.');
-    }
-    // Extract candidate's experience using regex
-    const experienceRegex = /EXPÉRIENCE PROFESSIONNELLE:\s*([\s\S]*?)(?=(FORMATION:|$))/;
-    const experienceMatch = extractedText.match(experienceRegex);
-    console.log('Experience match:', experienceMatch);
-
-    let experience = "Aucune expérience professionnelle trouvée";
-    if (experienceMatch && experienceMatch[1]) {
-      experience = experienceMatch[1];
-      console.log('Experience:', experience);
-    } else {
-      console.log('No experience found.');
-    }
-    // Extract languages from the text using regex
-    const languageRegex = /LANGUES:\s*([\s\S]*?)(?=(COMPÉTENCES:|CERTIFICATIONS:|$))/;
-    const languageMatch = extractedText.match(languageRegex);
-    console.log('Language match:', languageMatch);
-
-    let languages = "Aucune langue trouvée";
-    if (languageMatch && languageMatch[1]) {
-      languages = languageMatch[1].trim().split('\n').map(lang => lang.trim()).join(', ');
-      console.log('Languages:', languages);
-    } else {
-      console.log('No languages found.');
-    }
-    // Create a parsedDetails object containing extracted information
-    const parsedDetails = {
-      jobPosition: 'Stage',
-      candidateName: candidateName,
-      candidateEmail: candidateEmail,
-      candidateNum: candidateNum,
-      competences: competences,
-      postalAddress: postalAddress,
-      educationBackground: educationBackground,
-      certifications: certifications,
-      skills:skills,
-      experience:experience,
-      languages:languages
-
-    };
-    console.log('Parsed details:', parsedDetails);
-    return parsedDetails;
+  async parseResume(base64File: string): Promise<any>{
+    try {
+      const response = await fetch('https://eu-rest.resumeparsing.com/v10/parser/resume', { 
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json', 'Sovren-AccountId': '17097504', 'Sovren-ServiceKey': 'i8Stm46FEsltKLqQ2VNz1MzhCnlHORAYnOUO/dP7'},
+        body: JSON.stringify({ DocumentAsBase64String: base64File, DocumentLastModified: (new Date()).toISOString().substring(0, 10)}) }) 
+        const data = await response.json()  
+        return data.Value?.ResumeData; } 
+        catch (error) { console.log(`error when parseResume: ${error}`); return "Something went wrong";    }
   }
   async readURL(input: any, user: AppUser, workspace: any) {
     if (input.files && input.files[0]) {
@@ -247,7 +111,7 @@ export class CvParserService {
       candidateEmail: parsedResume.candidateEmail,
       originalCV: file.name,
       idworkspace: userWorkspace.idWorkspace,
-      designationStatus: 'En attente',
+      designationStatus: '',
       designationTemplate: 'Modèle 1'
     };
   
@@ -267,6 +131,11 @@ export class CvParserService {
   }
   async deleteCV(cvId: number): Promise<void> {
     try {
+      const shouldDelete = window.confirm('Êtes-vous sûr de vouloir supprimer ce CV ?');
+      if (!shouldDelete) {
+        return; 
+      }
+  
       const { data, error } = await this.supabase.from('CV').delete().eq('id_CV', cvId);
       if (error) {
         console.error('Error deleting CV from Supabase:', error);
@@ -279,30 +148,8 @@ export class CvParserService {
       throw error;
     }
   }
-  async extractTextFromPDF(pdfData: Uint8Array): Promise<string> {
-    try {
-      console.log('Extracting text from PDF...');
-      const pdf = await pdfjsLib.getDocument({ data: pdfData }).promise;
-      let text = '';
-      for (let pageNum = 1; pageNum <= pdf.numPages; pageNum++) {
-        const page = await pdf.getPage(pageNum);
-        const pageText = await page.getTextContent();
-        pageText.items.forEach(item => {
-          if ('str' in item) {
-            text += (item as any).str + ' ';
-          }
-        });
-      }
-      console.log('Extracted text from PDF:', text); 
-      return text;
-    } catch (error) {
-      console.error('Error extracting text from PDF:', error); 
-      throw error;
-    }
-  }
   
-
-  
+ 
   async parseResumeAndAddCV(base64File: string): Promise<any> {
     try {
       const parsedResume = await this.parseResume(base64File); 
@@ -329,4 +176,80 @@ export class CvParserService {
         throw error;
     }
   }
+
+fromSovren = async (resume: Resume, data: any) => {
+    resume.CandidateDetails.FirstName =
+        data.ContactInformation?.CandidateName?.GivenName || "";
+    resume.CandidateDetails.LastName =
+        data.ContactInformation?.CandidateName?.FamilyName || "";
+    resume.CandidateDetails.Email =
+        data.ContactInformation?.EmailAddresses?.[0]?.toString() || "";
+    resume.CandidateDetails.telephone =
+        data.ContactInformation?.Telephones?.[0]?.Normalized || "";
+    resume.CandidateDetails.role = "";
+    resume.CandidateDetails.Anneesexperience =
+        (data.EmploymentHistory?.ExperienceSummary?.MonthsOfWorkExperience || " ") +
+        " months";
+    data.Education &&
+        data.Education.EducationDetails &&
+        data.Education.EducationDetails.map((edu: any) =>
+            resume.Educations.Education.push({
+              Nom_ecole: edu?.SchoolName?.Normalized || "",
+                Diplome: edu?.Degree?.Type || "",
+                VilleE: "",
+                DatedebutF: edu.LastEducationDate
+                    ? edu.LastEducationDate.isCurrentDate
+                        ? "Present"
+                        : edu.LastEducationDate.Date
+                    : "",
+                DatefinF: edu.LastEducationDate
+                    ? edu.LastEducationDate.isCurrentDate
+                        ? "Present"
+                        : edu.LastEducationDate.Date
+                    : "",
+            })
+        );
+    data.EmploymentHistory &&
+        data.EmploymentHistory.Positions &&
+        data.EmploymentHistory.Positions.map((pos: any) =>
+            resume.historiques.Position.push({
+              Nomentreprise: pos?.Employer?.Name?.Normalized || "",
+              Intituleposte: pos?.JopTitle?.Normalized || "",
+              Datedebut: pos?.StartDate?.Date || "",
+              Datefin: pos.isCurrent ? "Present" : pos?.EndDate?.Date || "",
+              Description: pos?.Description || "",
+            })
+        );
+    data.Certifications &&
+        data.Certifications.map((cer: any) =>
+            resume.certifications.Certification.push({
+              titre_certificat: cer?.Name || "",
+              DateCert: "",
+            })
+        );
+    data.LanguageCompetencies &&
+        data.LanguageCompetencies.map((lan: any) =>
+            resume.Langues.Langue.push({
+              titre_langue: lan?.Language || "",
+              niveaulang: "Proficiant/Fluent",
+            })
+        );
+    resume.Competences.TopSkills = [...this.getTopSkills(data.SkillsData[0])];
+  
+
+  };
+  getTopSkills = (data: any) => {
+    const skills =
+      data?.Taxonomies?.flatMap((taxonomy: any) =>
+        taxonomy?.SubTaxonomies?.flatMap(
+          (subTaxonomy: any) => subTaxonomy?.Skills
+        )
+      )
+        ?.filter((skill: any) => skill?.ExistsInText)
+        ?.sort((a: any, b: any) => b?.PercentOfOverall - a?.PercentOfOverall)
+        ?.slice(0, 4)
+        ?.map((skill: any) => ({ Name: skill?.Name })) || [];
+    return skills as Competence[];
+  };
+
 }
